+++ WebClient/src/app/models/user.ts
@@ -14,12 +14,16 @@ export class User {
   PaymentMethod: string | null;
   LastFourDigits: string | null;
   CardExpiryDate: string | null;
+  Subscribed: boolean;
+  DOB: string | null;
+  DateOfBirth: string | null;
 
   constructor(
       Id: number, FirstName: string, LastName: string, Email: string,
       Phone: string, Address: string, Country: string, Password: string | null, Verified: boolean,
       City: string = '', State: string = '', PostalCode: string = '',
-      PaymentMethod: string | null = null, LastFourDigits: string | null = null, CardExpiryDate: string | null = null
+      PaymentMethod: string | null = null, LastFourDigits: string | null = null, CardExpiryDate: string | null = null,
+      Subscribed: boolean = false, DOB: string | null = null
   ) {
       this.Id = Id;
       this.FirstName = FirstName;
@@ -36,5 +40,8 @@ export class User {
       this.PaymentMethod = PaymentMethod;
       this.LastFourDigits = LastFourDigits;
       this.CardExpiryDate = CardExpiryDate;
+      this.Subscribed = Subscribed;
+      this.DOB = DOB;
+      this.DateOfBirth = DOB;
   }
 }